refactor: extract screen-name sanitize into its own type
PostHogLogger's lastScreenName was a delegate to PostHogSDK and its
sanitize/strip helpers were called from PostHogSDK.screen() across a
class boundary. Move sanitize + helpers into a new
PostHogScreenNameSanitizer enum under ScreenViews/ and drop the
Logger.lastScreenName accessor entirely. PostHogSDK.lastScreenName is
now the single source of truth.
